// frontend/src/lib/api/generated/model/projectListItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Construction ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatus } from './projectStatus.ts';

export interface ProjectListItem {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  description?: string | null;
  client_id: string;
  site_address?: string | null;
  city?: string | null;
  latitude?: string | null;
  longitude?: string | null;
  status?: ProjectStatus;
  planned_start?: string | null;
  planned_end?: string | null;
  actual_start?: string | null;
  actual_end?: string | null;
  contract_value?: string | null;
  retention_pct?: string | null;
  project_manager_id?: string | null;
  id: string;
  code: string;
  created_at: string;
  updated_at: string;
  client_name?: string | null;
  project_manager_name?: string | null;
}